move llm usage info class to helpers
